renamed evaluation clickhouse file
